requirements.txt
langgraph
langchain
python-dotenv
GitPython
langchain-mistralai
langsmith
tree-sitter 
tree-sitter-language-pack